CheckPoint scripts made, Checkpoint objects, checkpoint V1
Created two scripts for checkpoints, placed checkpoints around level 1. I am going to fix some checkpoint bugs. One bug is if you get the checkpoint as a prefab, it will not work. The other is sometimes the player does not teleport but their health is reset (health reset is intended)