Checkpoint model created and placed, checkpoint script updated, checkpoint symbol added, Timer displayed, Checkpoint information made, new materials added, GameController updated
The model for the checkpoint has been made. Now users can easily tell when they are at a checkpoint

Script updated so that when the user hits a checkpoint, the checkpoint's information is displayed on the left of the screen

Timer added to the top of the screen

GameController has been updated in order to display the extra information mentioned above. KEY NOTE: GameController was also updated by me in order to keep track of the current checkpoint and its information. The MasterCheckpointController script (That I added) reports back to it.
MasterCheckpointController keeps the information of EACH checkpoint object with the use of the Checkpoint Controller that I also added

Child = CheckpointController
Parent = MasterCheckpointController
Parent reports back to GameController

I created new materials for the checkpoint model